Re-build billing update flow
