+++ core/apis/subscriptions.py
@@ -12,13 +12,8 @@ def checkout():
     user = user_service.get_by_id(session['profile']['internal_user_id'])
     checkout_success_url = url_for('subscription.success', _external=True)
     checkout_cancel_url = url_for('subscription.cancel', _external=True)
-    if user.subscription:
-        existing_subscription_id = user.subscription.id
-    else:
-        existing_subscription_id = None
     session_id = subscription_service.create_billing_update_session(user.customer_id, checkout_success_url,
-                                                                    checkout_cancel_url, existing_subscription_id)
-    print('SEssionID: ', session_id)
+                                                                    checkout_cancel_url)
     return jsonify({'session_id': session_id})
